third_party/WebKit/LayoutTests/shadow-dom/slotchange-event.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script src='resources/shadow-dom.js'></script>
<div id='d1'>
  <template data-mode='open' data-expose-as='d1_shadow'>
    <div id='d1-d1'>
      <template data-mode='open' data-expose-as='d1_d1_shadow'>
        <slot name='d1-d1-s1'></slot>
      </template>
      <slot name='d1-s1' slot='d1-d1-s1'></slot>
    </div>
  </template>
  <div id='d2' slot='d1-s1'></div>
</div>
<script>
'use strict';
convertTemplatesToShadowRootsWithin(d1);
removeWhiteSpaceOnlyTextNodes(d1);

async_test((test) => {
  var count_for_d1_s1 = 0;
  var count_for_d1_d1_s1 = 0;

  const d1_s1 = d1_shadow.querySelector('slot');
  const d1_d1_s1 = d1_d1_shadow.querySelector('slot');

  assert_array_equals(d1_s1.assignedNodes(), [d2]);
  assert_array_equals(d1_s1.assignedNodes({'flatten': true}), [d2]);

  assert_array_equals(d1_d1_s1.assignedNodes(), [d1_s1]);
  assert_array_equals(d1_d1_s1.assignedNodes({'flatten': true}), [d2]);

  d1_s1.addEventListener('slotchange', (e) => {
    test.step(() => {
      count_for_d1_s1++;
      if (count_for_d1_s1 == 1) {
        assert_equals(e.target, d1_s1);
        assert_array_equals(d1_s1.assignedNodes(), [d2, d3]);
        assert_array_equals(d1_s1.assignedNodes({'flatten': true}), [d2, d3]);
        assert_array_equals(e.scoped, true);
      } else if (count_for_d1_s1 == 2) {
        assert_equals(e.target, d1_s1);
        assert_array_equals(d1_s1.assignedNodes(), [d2]);
        assert_array_equals(d1_s1.assignedNodes({'flatten': true}), [d2]);
        assert_array_equals(e.scoped, true);
        test.done();
      }
    });
  });

  d1_d1_s1.addEventListener('slotchange', (e) => {
    test.step(() => {
      count_for_d1_d1_s1++;
      if (count_for_d1_d1_s1 == 1) {
        assert_equals(e.target, d1_d1_s1);
        assert_array_equals(d1_d1_s1.assignedNodes(), [d1_s1]);
        assert_array_equals(d1_d1_s1.assignedNodes({'flatten': true}), [d2, d3]);
        assert_array_equals(e.scoped, true);
      } else if (count_for_d1_d1_s1 == 2){
        assert_equals(e.target, d1_d1_s1);
        assert_array_equals(d1_d1_s1.assignedNodes(), [d1_s1]);
        assert_array_equals(d1_d1_s1.assignedNodes({'flatten': true}), [d2]);
        assert_array_equals(e.scoped, true);
        test.done();
      }
    });
  });

  const d3 = document.createElement('div');
  d3.setAttribute('id', 'd3');
  d3.setAttribute('slot', 'd1-s1');
  d1.appendChild(d3);
  // Different microtask.
  setTimeout(function(){
    d3.remove();
  }, 0);

}, "slotchange event should be dispatched for a slot at most once per microtask.");
</script>
